backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-genai>=0.8.0
python-dotenv==1.0.1
pydantic==2.6.4
httpx==0.27.0